Modify workspace group manager

+++ src/spaceone/identity/manager/workspace_group_manager.py
@@ -6,6 +6,7 @@ from spaceone.core.error import ERROR_INVALID_PARAMETER, ERROR_NOT_FOUND
 from spaceone.core.manager import BaseManager
 
 from spaceone.identity.manager.role_binding_manager import RoleBindingManager
+from spaceone.identity.manager.workspace_manager import WorkspaceManager
 from spaceone.identity.model.workspace_group.database import WorkspaceGroup
 
 _LOGGER = logging.getLogger(__name__)
@@ -43,22 +44,34 @@ class WorkspaceGroupManager(BaseManager):
         return workspace_group_vo.update(params)
 
     def delete_workspace_group_by_vo(self, workspace_group_vo: WorkspaceGroup) -> None:
-        user_ids = [user["user_id"] for user in workspace_group_vo.users]
-        rb_vos = self.rb_mgr.filter_role_bindings(
-            user_id=user_ids,
+        if workspace_group_vo.users:
+            user_ids = [user["user_id"] for user in workspace_group_vo.users]
+            rb_vos = self.rb_mgr.filter_role_bindings(
+                user_id=user_ids,
+                workspace_group_id=workspace_group_vo.workspace_group_id,
+                domain_id=workspace_group_vo.domain_id,
+            )
+
+            if rb_vos.count() > 0:
+                _LOGGER.debug(
+                    f"[delete_workspace_group_by_vo] Delete role bindings count with {workspace_group_vo.users}: {rb_vos.count()}"
+                )
+                for rb_vo in rb_vos:
+                    _LOGGER.debug(
+                        f"[delete_workspace_group_by_vo] Delete role binding info: {rb_vo.to_dict()}"
+                    )
+                    rb_vo.delete()
+
+        workspace_mgr = WorkspaceManager()
+        workspace_vos = workspace_mgr.filter_workspaces(
             workspace_group_id=workspace_group_vo.workspace_group_id,
             domain_id=workspace_group_vo.domain_id,
         )
-
-        if rb_vos.count() > 0:
-            _LOGGER.debug(
-                f"[delete_workspace_group_by_vo] Delete role bindings count with {workspace_group_vo.users}: {rb_vos.count()}"
-            )
-            for rb_vo in rb_vos:
-                _LOGGER.debug(
-                    f"[delete_workspace_group_by_vo] Delete role binding info: {rb_vo.to_dict()}"
-                )
-                rb_vo.delete()
+        for workspace_vo in workspace_vos:
+            params = {
+                "workspace_group_id": None,
+            }
+            workspace_mgr.update_workspace_by_vo(params, workspace_vo)
 
         workspace_group_vo.delete()
